docker-greenplum/docker-compose/addons/pyclient/requirements.txt
Flask
confluent-kafka
kafka-python
psycopg2-binary
requests
